+++ src/coordinator/tests/TimerTest.php
@@ -113,6 +113,38 @@ class TimerTest extends TestCase
         $this->assertTrue(true);
     }
 
+    public function testUntil()
+    {
+        $this->wait(function () {
+            $id = 0;
+            $timer = new Timer();
+            $identifier = uniqid();
+            $timer->until(function () use (&$id) {
+                ++$id;
+            }, $identifier);
+
+            $this->assertSame(0, $id);
+            CoordinatorManager::until($identifier)->resume();
+            $this->assertSame(1, $id);
+        });
+    }
+
+    public function testUntilWhenClear()
+    {
+        $this->wait(function () {
+            $id = 0;
+            $timer = new Timer();
+            $identifier = uniqid();
+            $ret = $timer->until(function () use (&$id) {
+                ++$id;
+            }, $identifier);
+            $timer->clear($ret);
+            $this->assertSame(0, $id);
+            CoordinatorManager::until($identifier)->resume();
+            $this->assertSame(0, $id);
+        });
+    }
+
     private function wait(Closure $closure)
     {
         $waiter = new Waiter();