+++ src/database-pgsql/src/Schema/Grammars/PostgresGrammar.php
@@ -188,16 +188,15 @@ class PostgresGrammar extends Grammar
     {
         $language = $command->language ?: 'english';
 
-        if (count($command->columns) > 1) {
-            throw new RuntimeException('The PostgreSQL driver does not support fulltext index creation using multiple columns.');
-        }
+        $columns = array_map(function ($column) use ($language) {
+            return "to_tsvector({$this->quoteString($language)}, {$this->wrap($column)})";
+        }, $command->columns);
 
         return sprintf(
-            'create index %s on %s using gin (to_tsvector(%s, %s))',
+            'create index %s on %s using gin ((%s))',
             $this->wrap($command->index),
             $this->wrapTable($blueprint),
-            $this->quoteString($language),
-            $this->wrap($command->columns[0])
+            implode(' || ', $columns)
         );
     }